backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.15.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.2.0
Werkzeug==3.0.1